Create BoxHelper and FilesHelper
